binutils-pru: Disable libctf and sim
We don't need libctf for PRU. Simulator is only really useful for
regression-testing the toolchain, so no point in keeping it in the
Debian package.

This should fix simultaneous installation of Binutils packages for
different targets (PRU, AVR, MSP430).

Also remove bfd plugin directory. It's not needed. And it conflicts with
binutils packages for other targets.

Signed-off-by: Dimitar Dimitrov <dimitar@dinux.eu>